requirements.txt
fastapi
uvicorn[standard]
openai>=1.0
tavily-python
requests
python-dotenv
websockets
